Remove incorrect cast.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1194 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Master.java
@@ -191,7 +191,7 @@ nextRecord() throws IOException {
 			if (last == null && defaultTTL < 0)
 				throw st.exception("missing TTL");
 			else if (defaultTTL >= 0)
-				ttl = (int) defaultTTL;
+				ttl = defaultTTL;
 			else
 				ttl = last.getTTL();
 		}
